// src/components/SearchForm/api/airports/SearchAirport.ts
export interface SearchAirport {
  airlineCount: string;
  cityCode: string;
  countryCode: string;
  countryName: string;
  createdAt: string;
  displayLocation: string;
  displayName: string;
  flightsWeight: string;
  geopoint: string;
  iataCode: string;
  id: string;
  latitude: string;
  longitude: string;
  name: string;
  region: string;
  regionCode: string;
  routeCount: string;
  searchText: string;
  timezone: string;
  timeDifference: string;
  updatedAt: string;
  type?: string;
}